Re-organize web folders to better suit next.js.
