+++ proto/run_protoc.sh
@@ -2,7 +2,7 @@
 # date by running this.
 
 protoc --plugin="./web/node_modules/.bin/protoc-gen-ts_proto" \
-  --ts_proto_out="./web/app/src/lib" \
+  --ts_proto_out="./web/lib" \
   --go_out="./proto/go" \
   "./proto/bright_stars.proto" \
   "./proto/asteroids.proto"
