+++ web/lib/proto/asteroids.ts
