+++ web/lib/proto/bright_stars.ts
