+++ web/tsconfig.json
@@ -1,12 +1,18 @@
 {
   "compilerOptions": {
     "target": "ES2017",
+    "baseUrl": ".",
     "lib": [
       "dom",
       "dom.iterable",
       "esnext",
       "webworker"
     ],
+    "paths": {
+      "@/*": [
+        "*"
+      ]
+    },
     "allowJs": false,
     "skipLibCheck": true,
     "strict": false,
